Add IntSettingInput
